* Load testing settings only when DEBUG is true. * Require local and secrets settings files.
Previously, an exception from badly importing local.py and secrets.py
was caught but nothing was done to bail out. At that time, both files
were thought to be optional to override settings in __init__.py at the
administrator's discretion. Now, they are required to exist on a
per-server basis.