Fix problem with deleted and end_dated elements appearing in project list.
only_current = Q(end_date=None) | Q(end_date__gt=timezone.now())
This was not working as intended. There must be a bug. It's difficult
to reproduce but it doesn't always work. Instead it needs to be built up from
an empty Q(). This seems to fix it for now.

	modified:   api/project.py