BIG refactor of allocation/time usage
Separated concerns of 'who to suspend' Vs. auditing/monitoring the usage on a
particular account. The result is that counting up time on an instance is
more straightforward and the logged output produced can be converted to CSV.

NOTE: Users will NOT be happy because this is the first time that we MULTIPLY
CPU Cores by the active time used.