Removed size limit to prevent multiple rotates/day
If size is low, e.g. 100M and the file grows GB/day, it will rotate the log multiple times and could drop logs off the end faster than we wish to do so.  If this configuration is not preferred, we could try to put the size back in, but bump size up to 500M or so.