Revert "Avoid failure on empty database"
This reverts commit 5cb853387f7ce3e6b2dac28e266c4b48a60be5ab.

We don't actually want to handle this exception. This database model is
depended on in other places in the app. If this exception gets thrown, then
there are bigger issues like not having the base models we need. From an empty
database running the migrations (specifically
atmosphere/core/migrations/0019_set_default_status_type.py) would avoid this.