Fix nginx proxy pass unable to resolve localhost
Proxy_pass defines a uri for some proxied resource. If the uri contains a
variable nginx applies special rules. A resolver must be defined. We use
google's 8.8.8.8 dns. We used to use SERVER_NAME in the uri for the
resource, this works for productions servers with dns entries but fails for
local machines.

This solution removed the variables that were unnecessary and removed any
hostname that would need to be resolved.