Problem: New versions of an application, repli-...
-cated across providers were an unexpected edge case. Because the image
ID is *not* the same image that created the *application UUID*, no app
kwargs were passed down, potentially allowing this version to create a
new image. Instead, if REPLICATION_PROVIDER_LOCATION is set (redefined
'set' as getattr, not hasattr) attempt to find a similar image that
already exists in the stack, and use its application to define yourself.