Remove favicon branding when rendering 404 pages
With the adoption of Atmosphere growing, it is best to have no or
generic error/not-found/404 pages. There should be enough "styling"
that the web server does not appear to be correctly configured.
However, allowing for a favicon complicates the theming efforts,
overall. Here we offer "less" to gain the ability to focus on more.

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1536