Problem: Missing UserAllocationSources (and more!)
*Problem 1*

Users reported being unable to launch instances because they had no
allocation sources. On investigation it turned that the
`AllocationSources` with their names did exist, but not the
`UserAllocationSources` which were used to link their `AtmosphereUser`
accounts to the allocation sources.

We also found that the `user_allocation_source_created` events which
should trigger the creation of the `UserAllocationSources` did exist -
so that's not the problem.

The process for creating a `UserAllocationSource` relies on an
`AtmosphereUser` already existing, and happens as part of creating an
identity on a provider - in the
`service.accounts.openstack_manager.AccountDriver#create_account`.
Normally this happens when the user logs in and we know that the
`AtmosphereUser` exists.

We found another another place where `create_account` is called.

The `add_new_accounts.py` script is called by a cron job, looks up
new LDAP accounts with Atmosphere access, and creates identities on a
provider using `create_account`. It doesn't create an `AtmosphereUser`
before it calls `create_account`.

As a result `create_account` would fail when creating the
`UserAllocationSource`, but not before it creates an `AllocationSource`
matching the user name. Any calls to `create_account` also check if an
`AllocationSource` exists with the same name as the username so that it
does not create duplicate `AllocationSources`. It also does not try to
create (the now missing) `UserAllocationSource`.

This is how we ended up with an `AllocationSource` for a username as
well as an `user_allocation_source_created` event, but no
`UserAllocationSource`.

Solution: In `add_new_accounts.py`, before calling `create_account`
ensure that the AtmosphereUser exists by calling
`core.models.Group.create_usergroup`

Also add a 'safety valve'/'eject button' in
`service.accounts.openstack_manager.AccountDriver#create_account` to
throw an exception if is called with a username that is not present
in the system.

*Problem 2*

In `create_account` these allocation sources for usernames are also
being created on Jetstream. This is not good. Fortunately they are
being deleted as soon as they are created because of the TAS
'mirroring' logic, but still...

Solution: Created an AllocationSourcePluginManager

There are two plugins to start with:

- `jetstream.plugins.allocation_source.JetstreamAllocationSource`
- `cyverse_allocation.plugins.allocation_source.FlexibleAllocationSource`

Currently the only required method for a plugin of this type is:
`ensure_user_allocation_source` and it should return a boolean.

This will ensure that the user has a valid allocation source,
potentially creating it in the case of the `FlexibleAllocationSource`.

I still have to implement the logic for `JetstreamAllocationSource`, and
it just returns `True` for now.