Remove bad validation
This method is not validating. If it were it should override:

    # http://www.django-rest-framework.org/api-guide/serializers/#object-level-validation
    def validate(self, data):

This create if it doesn't exist belongs in a migration not in validation
logic.

As a side note (or a practical reason to avoid this). After removing this, I
found a validation issue where we were requiring "created_by" /despite/ our
`create` method on the serializer making it optional. So with this change I
also made "created_by" required=False.