Prefer filtering in python to database queries with queryset.filter
The view that renders an Identity can prefetch the Identity's credential set.
This means that an Identity won't have to execute a db query every time it
needs to lookup a credential. If there is no prefetch, then this will be
slightly slower than the existing solution. However, there is a dramatic
speedup (in aggregate) when prefetch is used.

Remove unnecessary abstraction `openstack_key`
```diff
     def get_key(self):
-        return self.openstack_key()
-
-    def openstack_key(self):
+        creds = self.get_credentials()
         return "Username: %s, Project:%s" % (
-            self.get_credential('key'),
-            self.project_name())
```

Rewrite project_name to be less verbose
```
+    def project_name(self, creds=None):
+        creds = creds or self.get_credentials()
+        return creds.get("ex_project_name", False) or \
+               creds.get("ex_tenant_name", False)  or \
+               creds.get("project_name", False)    or \
+               creds.get("tenant_name", False)     or \
+               ""
```

Remove the filter (db query), and iterate over all creds that are prepopulated
```diff
     def get_credential(self, key):
-        cred = self.credential_set.filter(key=key)
-        return cred[0].value if cred else None
+        for cred in self.credential_set.all():
+            if cred.key == key:
+                return cred.value
+        return None
```

In the Hotfix portion, this method needed to use the logic of project_name, but
project_name needed the logic of this method. So project_name can operate with
a passed dictionary of credentials and the dependency is resolved.
```diff
     def get_credentials(self):
         cred_dict = {}
         for cred in self.credential_set.all():
             cred_dict[cred.key] = cred.value
+
         # Hotfix to avoid errors in rtwo+OpenStack
         if 'ex_tenant_name' not in cred_dict:
-            cred_dict['ex_tenant_name'] = self.project_name()
+            cred_dict['ex_tenant_name'] = self.project_name(creds=cred_dict)
         return cred_dict
```