Problem: Emulate is broken
While fixing the emulate endpoint in PR #464, I broke it in a different
way. I didn't fully understand how the emulate process worked. I assumed
the `username` and the `user` were interchangeable based on the other
examples of `get_or_create_token`. But in this case the `username`
corresponded to a different user (the user to emulate), and not to the
logged-in user. As a consequence the user ended up emulating themselves.

Solution: Look up the user to emulate and pass that into the
`get_or_create_token` function.