Problem: Creating volume (for project) fails
URL: `/api/v2/project_volumes`

Error:

```
django.db.models.fields.related_descriptors:RelatedObjectDoesNotExist: Volume has no instance_source.
```

Background: The refactoring of the project-volumes relationship meant
that the many-to-many table doesn't exist anymore. The way the new code
was structured meant that a new volume was created, instead of setting
the `project` attribute on an existing volume.

We only caught it because the volume blew up due to a missing attribute
- `instance_source`.

Solution: Added an explicit `create` method on `ProjectVolumeSerializer`
which sets the `project` attribute on the volume

This overrides the default creation behavior which only made sense in
the prior arrangement.