Mark private images as "shared" => you + admin
With changes made to handle group ownership, images that were "private"
are considered "shared"; the explanation is that they're shared with a
user (you) and the admin - always. So a private or shared with specific
user image are both treated as "shared" within the glance catalog (I
think).

This is noted and captured here as a fix for issue related to imaging
encountered during a verification of "private" images, and this updated
semantic is required.