Update quota for all user's identities
This is a hack.

Problem:
In j7m, users don't make resource requests against specific
providers, yet each request is tied to a provider (identity). So admins in j7m
are approving quota essentially for one of the user's identities.

Solution:
In the action for changing quota (specific to resource requests)
update the quota for all the user's identities.