Split resource requests into user and admin api
Resource requests no longer have an associated cloud identity. Resource
requests used to inherit from BaseRequest (like MachineRequest). In order to
remove the identity w/o affecting MachineRequest, I copied/reworked the
relevant bits, so that the model ResourceRequest extends Model.

There are now two endpoints:

1) api/v2/resource-requests
2) api/v2/admin/resource-requests

The first API returns the user's requests and has custom permissions for a
user. They are only allowed to update the status of their requests to "closed".

The second API returns all users requests, it is only viewable by the admin.
The admin can update the status to approve or deny.

This separation of apis makes it trivial to implement permissions correctly.
One endpoint has all permissions the other has a subset.

We've encountered this issue in the past--how we should design the API to
support an admin and user view. The problem with using query paramaters
(&admin=true) is that Django permissions are harder to set and the troposphere
stores interpret query parameters as filters on the resources, so that would
require redesign.

Both apis have tests written to ensure the refactor should work.