Fix old requirment that resource request is associated with identity
This fixes a failing test when the resource request email is sent. The admin
url is supplied and passed by troposphere. It's optional, the default url is
constructed from the user's selected identity. Because the test user has no identities, the test failed.

There is no longer any reason to try and construct this admin_url in the
template email from an identity, so it has been removed.

The tests have been updated to reflect that troposphere provides an admin_url. The email template supports an optional admin_url if it's not supplied.