Fix unbound local variable username
username was throwing unbound because request.session.get(...) was throwing an
attribute error. The attribute error was thrown because the request object
doesn't have a session attribute in a test environment, because the test
environment doesn't include SessionMiddleware.

So its an error to assume that username would be given a value, that was fixed
by username = None, and its equally bad to catch all exceptions, that was
fixed