Fix undefined variable SECRET_KEY, remove unused variable
The linter would complain, because SECRET_KEY was never imported in the file.
We were doing an unnecessary hack of mutating the variable in the current
namespace to see SECRET_KEY. It turns out that the last use of SECRED_SEED was
a while ago, so this code which caused the lint err can be nuked.