Lint local.py
A #noqa line was added because Dodgy thought we were hardcoding a password

ORG_NAME,SITE_NAME, and SITE_TITLE can be undefined in the variables.ini,
which would result in those variable never being declared, even those later in
settings/local.py expected them to exist. In the future we will not conditionally
define variables!