Raise separate error for blacklisted allocations
Problem: when allocation name is blacklisted, user gets cryptic error when
launching
Solution: Give them a precise error

This is sort of an edge case but warranted. Troposphere no longer performs
allocation validation before launching, it defers to the backend to raise
exceptions and thus trusts the backend to have good errors.