Remove unnecessary ENFORCING check
Problem: Testing allocation logic requires setting ENFORCING=true, which is dangerous
Solution: Remove ENFORCING from the equation

This check prevented allocation enforcement logic taking place when
ENFORCING=false. That means that when ENFORCING=false (in the case of a dev
environment), that we would not do checking of your allocation when instances
were launched. This is a part of the logic we definitely want to be included
in dev environements, and we shouldn't have to enable ENFORCING=true do to
that, which iss verrryy bad outside of production.