Fix unshelved instance failing to get fixed ip
I changed the name of add_fixed_ip, because we are taking a quite different
approach here. We delete all an instance's ports, and then instruct nova to
attach the instance to a subnet. Nova takes care of creating a port.

This fixes an outstanding issue in nova[0]. Our original approach is to keep
ports around and then just update the fixed ips. However, when an instance and
its port are created, they each have an availability zone, when we unshelve an
instance its AZ may change. If it does change, openstack fails to update the
device owner AZ of the port and attaching a fixed ip fails because no ports
exist in the AZ.

We also used to exit from add_fixed_ip if the instance already had a fixed ip.
However, we can no longer do this. Instances affected by this shelving bug,
will never have their fixed ips removed, so if we exited early. Instances that
need a new port (on the new AZ) would not receive a new port.

https://bugs.launchpad.net/nova/+bug/1759924[0]