Revert "Fix unshelved instance failing to get fixed ip"
We are no longer removing ports. After upgrading Marana to Pike we found that
ports which were attached post boot, would never successfully attach. A
workaround was discovered to enable allow-hotplug for the given interface
 in cloud-init, but it was deemed more complicated than simply preserving the
 ports.

We were deleting/recreating the ports in the first place, because of a bug
which affected j7m https://bugs.launchpad.net/nova/+bug/1759924.

This reverts commit af002eb03892374eda4654f9218bb6f8f3ad9973.