Fix quota update causing index out of range
Problem:
Updating a users quota, would trigger an index out of range error

Solution:
Use a different api for fetching the users tenant id

In the act of changing quota, we require a tenant id for the user. The original api
which fetched this id, did so through a side effect of establishing a
connection to the nova compute api. This side effect changed when marana was
upgraded to use the 2.1 compute api.

We are no longer using this method in rtwo `_get_tenant_id`.