Fix monitor_machines_for in the presence of inactive clouds
Problem:
AttributeError: 'NoneType' object has no attribute 'get_image_members'

Solution:
Remove irrelevant code

During monitor_machines_for we fetch the latest images from the cloud and sync
those with our database. Part of that syncing is creating the Membership
relations which affect who can see an atmosphere image (more correctly a
provider machine). There was a bug in this process.

For each image we convert to a provider machine, we check if a machine request
exists on any provider. If one does exist and the provider of the request
doesn't equal the provider of the machine we're syncing, then we copy the image
members of that original provider onto this provider machine. The error is that
we never check if that original provider is active (i.e still running). If it's
not active, then we cannot fetch the image members. This occurred after the
Tucson cloud was made inactive.

We do not need to be copying image members during monitor_machines_for in the
first place. That should be part of image synchronization between providers.