Fix mount failure despite attach success
Problem:
After what appears like a successful volume attach, the volume would not be mounted in the instance

Solution:
Wait to mount until the volume reports "in-use"

When attaching a volume against Openstack, atmosphere would check if the
volume was in attaching. If it was not in that state, it would exit without
failure. The chain would continue and the ansible would attempt to mount,
despite the volume not being ready.

Normally the volume moves immediately to "attaching" however occasionally it
would move first to "reserved".

The code has been altered to wait until the volume is "in-use", it will either
timeout trying or get there eventually. Additionally, if the volume doesn't
report the proper attachments it will raise an exception and not silently fail
as before.

It is possible that if the volume is attached elsewhere that this approach
will not work. Neither this fix not the original code was designed to handle
multiple volume attachments.