Remove incorrect check
This check is unnecessary in any case, but it should have checked that the
metadata key was present, instead it checks that the metadata value is truthy.
For example if `instance.extra.metadata == {}` that is sufficient, but that
would cause this check to fail. The check would fail for a mock instance.