On project delete, do not cascade to volumes and instances
Problem:
Deleting a project in api/v2 deletes end_dated instances/volumes and all
cascaded objects

Solution:
Set instance.project = NULL, volume.project = NULL when related project is deleted

Projects are not end-dated but deleted. Further, projects hold on to
instances, when the instances are end-dated. So a terminated instance will
still have a link to its project, and that project a link to that instance.
When a project is deleted, the delete cascades to all related objects that
depend on a project. The first to go are all the end-dated volumes/instances
in that project, then all the instance status histories of those instances.
Now those related models will not be deleted.