Test deleting a project does not delete instances
When an instance is end_dated it is not removed from a project. Because we
perform actual deletes on projects, this normally has a cascading delete
effect on instance or volumes that were enddated and still associated with a
project. We otherwise do not allow projects to be deleted that have active
instances or volumes.