After deleting, remove instance from cache
This is questionably worth doing. We're doing it because we still like
api.v1.instance which reads from the cache. So in theory by removing the
instance from the cache we'll get the right responses from api.v1. In practice
this isn't true because the entire api.v1 endpoint rests under a separate
django cache. That will be fixed later.