Fix test relying on certain settings
core.tests.test_machine.TestMachineMonitoring.test_monitoring_with_cyverse_validation
was failing because the test assumed that settings.BLACKLIST_METADATA_KEY was
set to atmo_image_exclude. The variables.ini.dist results in it being set to
the empty string. This change fixes the test to use the current settings,
rather than implicitly relying on settings to be defined in a given way.