Fix indirect import and unhappy linter
The linter was reporting:

    Fix [F401(pyflakes), None] core.serializers.fields.ModelRelatedField imported but unused

For the file api/v2/serializers/fields/base.py, because it was importing the
reference and not using it in that file. However, other files in turn imported
from this base file.

The solution is to import the class from package where it is defined. It was
defined in the core.serializers.fields package.