Fix TACC_API_READ_TIMEOUT to raise a TASAPIException
When validation occurs in:
jetstream.plugins.auth.validation.XsedeProjectRequired.validate_user

It calls out to the tas api for a user's allocations. It falls back to the
database, if a TASAPIException occurs. We want it to fallback when the tas api
is not responding. So we need to raise a TASAPIException when a timeout
occurs.