Remove expensive cloud check
This check queries all the instances in the cloud. It blocks requests to
unshelve. If the query takes 30 sec, then the request to unshelve takes at
least 30 sec.

On IU, the query takes a minute. This is way too long for a single api
request. If 24 people tried to unshelve the entire app would be hung. nginx
was timing out for unshelving instances on IU because of how long this took,
with a Gateway Timeout Error.