Change driver to support making requests with a timeout
Several changes here that are all related. The original desired behavior for
the timeout feature was to ensure that api/v1/profile didn't take an
indefinite amount of time in user validation.

We originally accomplished that by putting a timeout of 5 seconds into all
calls to read from the tas api. This was wrong because there are some calls
which rightfully should take more time (like getting all projects).

Th purpose of this change is to add support for timeouts so that the one
place that needed it could use it w/o affecting all the other places.

This refactor allows you construct a driver with a specific timeout. That way
the one place (user validation) could construct a driver with a short timeout,
and handle a timeout exception appropriately.

This change introduces private methods _tacc_api_get and _tacc_api_post to the
driver so that it can provide a timeout in one place, and pass things like
tacc_password once. The tacc_api_get and tacc_api_post methods have been
changed to support timeout. They will eventually be incorporated into the
driver.